Add files via upload
In this program, I make use of my knowledge of PHP, HTML, CSS, JavaScript, and SQL to create a database that connects to my website to and the database. The website opens with an option for the user to enter their phone number, if the phone number is not in the database we will send an alert. Within the database, there is a table written in SQL holding the user's number, name, and address. After we check if the information is in the database the user is taken to a second page with the food truck order page where they have the option to order foods such as tacos and a drink and when the order is placed they will get their total and a confirmation alert.